Sort order, pagination work
